Authentication, Login, Pagination, Search, Users
"Implemented login screen with authentication using DummyJSON's auth API. User login is handled with a token and user information stored in the context API, while the token is also stored in localStorage. If the page is refreshed, the token is decrypted to retrieve the user data and update the state.

Designed the dashboard page with static data for now.

Developed the users page with pagination for listing users. Implemented a search functionality that dynamically updates the search, page, and pageSize query parameters as the page changes or a search is performed. Created URL linking to allow users to continue from the same point if they enter a query-based link. Implemented forms for user deletion and editing, which are instantly reflected to the user through API requests."